move NewWorker to worker.go
